L16 added possibility to create order and bucket, save orders in  database

+++ Homework/java-web/domain/src/main/java/com/konstantion/product/dto/CreationProductDto.java
@@ -1,4 +1,4 @@
 package com.konstantion.product.dto;
 
-public record CreationProductDto(String name, Integer price) {
+public record CreationProductDto(String name, Double price) {
 }
